pyproject.toml
[tool.poetry]
name = "internproject"
version = "0.1.0"
description = ""
authors = ["Parker Burchett <parker@tokemak.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
web3 = {extras = ["tester"], version = "^5.28.0"}
pandas = "^1.4.3"
pyarrow = "^11.0.0"
requests = "^2.28.1"
numpy = "^1.23.1"
scipy = "^1.8.1"
ruff = "^0.0.253"
black = {extras = ["d", "jupyter"], version = "^23.1.0"}
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
plotly = "^5.9.0"
jupyterlab = "^3.4.7"
nbconvert = "^7.3.0"


[tool.ruff]
line-length = 120
[tool.ruff.per-file-ignores]
'__init__.py' = ['F401', 'F403']

[tool.black]
line-length = 120
